back/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
SQLAlchemy==2.0.25
python-jose==3.3.0
passlib==1.7.4
email-validator==2.1.0.post1
